+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Added [Get-OGP][] Acknowledgements and License Info.
+
 
 ## [0.18.0] - 2025-11-23